feat(odyssey-react-mui): move illustration to it's own file
