+++ packages/odyssey-react-mui/src/labs/FileUpload.tsx
@@ -30,6 +30,7 @@ import {
   DesignTokens,
 } from "../OdysseyDesignTokensContext";
 import { FileUploadPreview } from "./FileUploadPreview";
+import { FileUploadIllustration } from "./FileUploadIllustration";
 
 export const fileUploadTypes = ["single", "multiple"] as const;
 export const fileUploadVariants = [
@@ -106,44 +107,6 @@ const ButtonAndInfoContainer = styled.div`
   align-items: center;
   justify-content: center;
 `;
-
-const UploadIllustrationContainer = styled("div", {
-  shouldForwardProp: (prop) => prop !== "odysseyDesignTokens",
-})<{
-  odysseyDesignTokens: DesignTokens;
-}>`
-  margin-block-end: ${({ odysseyDesignTokens }) =>
-    odysseyDesignTokens.Spacing2};
-  padding: ${({ odysseyDesignTokens }) => odysseyDesignTokens.Spacing3};
-  background-color: ${({ odysseyDesignTokens }) =>
-    odysseyDesignTokens.HueNeutral50};
-  border-radius: 50%;
-
-  svg {
-    display: flex;
-    width: ${({ odysseyDesignTokens }) => odysseyDesignTokens.Spacing8};
-    height: ${({ odysseyDesignTokens }) => odysseyDesignTokens.Spacing8};
-  }
-`;
-
-const UploadIllustration = () => {
-  const odysseyDesignTokens = useOdysseyDesignTokens();
-
-  return (
-    <UploadIllustrationContainer odysseyDesignTokens={odysseyDesignTokens}>
-      <svg viewBox="0 0 44 45" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path
-          d="M32.0763 11.001C29.3564 3.7855 21.6595 -0.565827 13.7765 0.726748C5.35441 2.10773 -0.676662 9.50714 0.0603005 17.8612C0.441865 22.1865 2.56458 25.9787 5.71703 28.614L8.28246 25.545C5.90122 23.5544 4.32811 20.7209 4.04483 17.5097C3.50262 11.3633 7.94433 5.73648 14.4238 4.67404C20.9164 3.60944 27.0806 7.52016 28.6895 13.5191C28.9239 14.3932 29.7162 15.001 30.6212 15.001H32.9114C36.8985 15.001 39.9997 18.0938 39.9997 21.7505C39.9997 24.3423 38.4576 26.6352 36.1259 27.7678L37.8736 31.3658C41.4737 29.6171 43.9997 25.9917 43.9997 21.7505C43.9997 15.7428 38.963 11.001 32.9114 11.001H32.0763Z"
-          fill="#E1E1E1"
-        />
-        <path
-          d="M23.9994 29.3277V44.5H19.9994V29.3289L14.4142 34.9141L11.5858 32.0857L19.7373 23.9342C20.9869 22.6845 23.0131 22.6845 24.2627 23.9342L32.4142 32.0857L29.5858 34.9141L23.9994 29.3277Z"
-          fill="#E1E1E1"
-        />
-      </svg>
-    </UploadIllustrationContainer>
-  );
-};
 
 export type FileUploadProps = {
   /**
@@ -275,7 +238,7 @@ const FileUpload = ({
           <InputContainer odysseyDesignTokens={odysseyDesignTokens}>
             <Input />
             <ButtonAndInfoContainer>
-              <UploadIllustration />
+              <FileUploadIllustration />
               <Support color="textSecondary">
                 Drag and drop files here or click to add files.
               </Support>
